feat: Add ability for command-line and environment override of service name
Command line options are -sk/--serviceKey
Environment varable is EDGEX_SERVICE_KEY

Includes support for the <profile> replacement text placeholder

closes #355

BREAKING CHANGE: Inserting preceding "-" when replacing `<profile>` in the service key has been removed so the use is more flexible.  The only service using the <profile> replacement text is app-service-configurable which will be updated to add the "-" in the initial service key.

Signed-off-by: lenny <leonard.goodell@intel.com>