feat: Optimize findMatchingFunction (#1071)
Sometimes the name of a newly added configurable pipeline function may start with one of the existing function names. For example, user is unaware of the existence of AddTags, implements function AddTagsFromDeviceResource and uses it in App Service configuration. In this case, sdk will resolve it as the function AddTags.
To make the result more accurate, the findMatchingFunction should try to find if there is a function with name identical to the target configuration function first. If no match, then iterate over all the functions to see if the target configuration function name starts with actual function name. If there are multiple matches then pick the one with the longest name.

Signed-off-by: Felix Ting <felix@iotechsys.com>