fix(webserver): fix a nil pointer crash when enabling the https web server
When enabling HTTPS, the web server would try to use dic, but it was never set in the NewWebServer() function.
This would result in a nil pointer panic when trying to access Vault.

Also, a tls.Config was created because ListenAndServeTLS requires the filename and path of a certificate and key but the raw certificate and key data
was retrieved from Vault. Leaving the cert and key as empty strings uses the certificate in tls.Config.

Fixes: #1315

Signed-off-by: Nick Yameen <nickyameen21@gmail.com>